CIF-2147: remove anonymous inner classes
